if loged in favourites and add event is disabled
