+++ src/navigation/Sidebar.js
@@ -9,19 +9,33 @@ import stats from "./stats.png";
 
 
 
-const Sidebar = () => {
+const Sidebar = (props) => {
     return(
         <div className="wrapper">
         <Nav variant="pills" defaultActiveKey="/home" className="flex-column">
         <Nav.Item >
         <Nav.Link exact as={NavLink} to="/"> <img src={planner} height="25"/><span>Wydarzenia</span></Nav.Link>
         </Nav.Item>
-        <Nav.Item>
-        <Nav.Link as={NavLink} to="/addEvent"> <img src={addNew} height="25"/><span>Nowe wydarzenie</span></Nav.Link>
-        </Nav.Item>
-        <Nav.Item>
-        <Nav.Link as={NavLink} to="/favourite"> <img src={favourite} height="25"/><span>Ulubione</span></Nav.Link>
-        </Nav.Item>
+        {
+             props.user ?
+            <Nav.Item>
+            <Nav.Link as={NavLink} to="/addEvent"> <img src={addNew} height="25"/><span>Nowe wydarzenie</span></Nav.Link>
+            </Nav.Item>
+            :
+            <Nav.Item>
+            <Nav.Link disabled as={NavLink} to="/addEvent"> <img src={addNew} height="25"/><span>Nowe wydarzenie</span></Nav.Link>
+            </Nav.Item>
+        }
+        {
+            props.user ?
+            <Nav.Item>
+                <Nav.Link as={NavLink} to="/favourite"> <img src={favourite} height="25"/><span>Ulubione</span></Nav.Link>
+            </Nav.Item>
+            :
+            <Nav.Item>
+                <Nav.Link disabled as={NavLink} to="/favourite"> <img src={favourite} height="25"/><span>Ulubione</span></Nav.Link>
+            </Nav.Item>
+        }
         <Nav.Item>
         <Nav.Link as={NavLink} to="/statistics"><img src={stats} height="25"/><span>Statystyki</span></Nav.Link>
         </Nav.Item>
